Refactor LifeInsuranceRepo, PropertyInsuranceRepo and VehicleInsuranceRepo
